Move airnode examples scripts to "src" folder
