+++ packages/airnode-examples/package.json
@@ -7,27 +7,27 @@
   "scripts": {
     "build": "yarn run clean && hardhat compile && yarn tsc --build tsconfig.json",
     "clean": "rimraf -rf *.tsbuildinfo ./dist *.tgz",
-    "choose-integration": "ts-node scripts/choose-integration.ts",
-    "create-airnode-config": "ts-node scripts/create-airnode-config.ts",
-    "create-airnode-secrets": "ts-node scripts/create-airnode-secrets.ts",
-    "create-aws-secrets": "ts-node scripts/create-aws-secrets.ts",
-    "deploy-airnode": "ts-node scripts/deploy-airnode.ts",
-    "deploy-requester": "ts-node scripts/deploy-requester.ts",
-    "deploy-rrp": "ts-node scripts/deploy-rrp.ts",
-    "derive-and-fund-sponsor-wallet": "ts-node scripts/derive-and-fund-sponsor-wallet.ts",
+    "choose-integration": "ts-node src/scripts/choose-integration.ts",
+    "create-airnode-config": "ts-node src/scripts/create-airnode-config.ts",
+    "create-airnode-secrets": "ts-node src/scripts/create-airnode-secrets.ts",
+    "create-aws-secrets": "ts-node src/scripts/create-aws-secrets.ts",
+    "deploy-airnode": "ts-node src/scripts/deploy-airnode.ts",
+    "deploy-requester": "ts-node src/scripts/deploy-requester.ts",
+    "deploy-rrp": "ts-node src/scripts/deploy-rrp.ts",
+    "derive-and-fund-sponsor-wallet": "ts-node src/scripts/derive-and-fund-sponsor-wallet.ts",
     "dev:generate-example-files": "ts-node dev-scripts/generate-example-files.ts",
     "eth-node": "hardhat node",
-    "make-request": "ts-node scripts/make-request.ts",
-    "make-withdrawal-request": "ts-node scripts/make-withdrawal-request.ts",
-    "print-account-details": "ts-node scripts/print-account-details.ts",
-    "rebuild-airnode-container": "ts-node scripts/rebuild-airnode-container.ts",
-    "rebuild-artifacts-container": "ts-node scripts/rebuild-artifacts-container.ts",
-    "rebuild-deployer-container": "ts-node scripts/rebuild-deployer-container.ts",
-    "remove-airnode": "ts-node scripts/remove-airnode.ts",
-    "run-airnode-locally": "ts-node scripts/run-airnode-locally.ts",
-    "sponsor-requester": "ts-node scripts/sponsor-requester.ts",
-    "stop-local-airnode": "ts-node scripts/stop-local-airnode.ts",
-    "test:coingecko-aws": "ts-node scripts/aws-test/coingecko-aws-test.ts",
+    "make-request": "ts-node src/scripts/make-request.ts",
+    "make-withdrawal-request": "ts-node src/scripts/make-withdrawal-request.ts",
+    "print-account-details": "ts-node src/scripts/print-account-details.ts",
+    "rebuild-airnode-container": "ts-node src/scripts/rebuild-airnode-container.ts",
+    "rebuild-artifacts-container": "ts-node src/scripts/rebuild-artifacts-container.ts",
+    "rebuild-deployer-container": "ts-node src/scripts/rebuild-deployer-container.ts",
+    "remove-airnode": "ts-node src/scripts/remove-airnode.ts",
+    "run-airnode-locally": "ts-node src/scripts/run-airnode-locally.ts",
+    "sponsor-requester": "ts-node src/scripts/sponsor-requester.ts",
+    "stop-local-airnode": "ts-node src/scripts/stop-local-airnode.ts",
+    "test:coingecko-aws": "ts-node src/scripts/aws-test/coingecko-aws-test.ts",
     "test": "jest --selectProjects unit",
     "test:e2e": "jest --selectProjects e2e"
   },
