--- packages/airnode-examples/scripts/rebuild-artifacts-container.ts
@@ -1,7 +1,0 @@
-import { runAndHandleErrors, runShellCommand } from '../src';
-
-const main = async () => {
-  runShellCommand(`yarn --cwd ../../ docker:build:artifacts`);
-};
-
-runAndHandleErrors(main);
