--- packages/airnode-examples/scripts/tsconfig.json
@@ -1,11 +1,0 @@
-{
-  "extends": "../tsconfig.json",
-  "compilerOptions": {
-    "composite": true,
-
-    "rootDir": "..",
-    "outDir": "../dist"
-  },
-  "references": [{ "path": "../src" }],
-  "include": ["./**/*.ts"]
-}
