+++ packages/airnode-examples/src/scripts/create-airnode-config.ts
@@ -1,10 +1,10 @@
-import { readIntegrationInfo, runAndHandleErrors } from '../src';
+import { readIntegrationInfo, runAndHandleErrors } from '../';
 
 const main = async () => {
   const integrationInfo = readIntegrationInfo();
   // Import the "create-config" file from the chosen integration. See the respective "create-config.ts" file for
   // details.
-  const createConfig = await import(`../integrations/${integrationInfo.integration}/create-config.ts`);
+  const createConfig = await import(`../../integrations/${integrationInfo.integration}/create-config.ts`);
   await createConfig.default();
 };
