+++ packages/airnode-examples/src/scripts/deploy-airnode.ts
@@ -1,12 +1,5 @@
 import { join } from 'path';
-import {
-  cliPrint,
-  isWindows,
-  readIntegrationInfo,
-  readPackageVersion,
-  runAndHandleErrors,
-  runShellCommand,
-} from '../src';
+import { cliPrint, isWindows, readIntegrationInfo, readPackageVersion, runAndHandleErrors, runShellCommand } from '../';
 
 const main = async () => {
   const integrationInfo = readIntegrationInfo();
@@ -21,15 +14,16 @@ const main = async () => {
   const packageVersion = readPackageVersion();
   const imageName = specificImage.length === 1 ? specificImage : `api3/airnode-deployer:${packageVersion}`;
 
-  const integrationPath = join(__dirname, '../integrations', integrationInfo.integration);
-  const secretsFilePath = join(__dirname, '../aws.env');
+  const integrationPath = join(__dirname, '../../integrations', integrationInfo.integration);
+  const secretsFilePath = join(__dirname, '../../aws.env');
   const deployCommand = [
     `docker run -it --rm`,
     isWindows() ? '' : `-e USER_ID=$(id -u) -e GROUP_ID=$(id -g)`,
     integrationInfo.airnodeType === 'aws' && `--env-file ${secretsFilePath}`,
     integrationInfo.airnodeType === 'gcp' && `-v "${integrationPath}/gcp.json:/app/gcp.json"`,
+    `-v ${integrationPath}:/app/config`,
     `-v ${integrationPath}:/app/output`,
-    `${imageName} remove -r output/receipt.json`,
+    `${imageName} deploy`,
   ]
     .filter(Boolean)
     .join(' ');
