+++ packages/airnode-examples/src/scripts/rebuild-artifacts-container.ts
@@ -1,0 +1,7 @@
+import { runAndHandleErrors, runShellCommand } from '../';
+
+const main = async () => {
+  runShellCommand(`yarn --cwd ../../../ docker:build:artifacts`);
+};
+
+runAndHandleErrors(main);
