+++ packages/airnode-examples/tsconfig.json
@@ -1,5 +1,5 @@
 {
   "extends": "../../tsconfig.json",
   "files": [],
-  "references": [{ "path": "./src" }, { "path": "./test" }, { "path": "./scripts" }, { "path": "./dev-scripts" }]
+  "references": [{ "path": "./src" }, { "path": "./test" }, { "path": "./dev-scripts" }]
 }
